requirements.txt
requests>=2.32.0
PyYAML>=6.0.1
rich>=13.7.0
python-dotenv>=1.0.1
pytest>=8.0.0
